Reduce memory usage for S3.
The problem was that for each S3TileStore we were instanciating a different
boto3.session.Session. This guy contains a cache for the whole AWS API
specification, including the doc. Roughly 3MB wasted by instance.

By using the boto3.client helper function, we use a global session, shared
between the connections. In tilecloud-chain, we use boto3.resource to
get the SQS queue and this guy uses the global session as well.